+++ opencascade-hs/cpp/hs_types.h
@@ -90,6 +90,8 @@ typedef void Poly_Triangulation;
 typedef void Poly_Triangle;
 typedef void BRepAdaptor_Curve;
 typedef int GeomAbs_CurveType;
+typedef void GeomConvert_BSplineCurveToBezierCurve;
+
 #define Handle(X) void
 #define ARRAY_1(X) void
 #else // __cplusplus