+++ opencascade-hs/src/OpenCascade/GeomConvert/Types.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.GeomConvert.Types
+( BSplineCurveToBezierCurve
+) where
+
+data BSplineCurveToBezierCurve
+